requirements.txt
# requirements.txt
discord.py
PayPaython-mobile
fastapi
uvicorn
httpx
pydantic
python-dotenv
requests
asyncio
json
os
time
shutil
pathlib
lief
beautifulsoup4
python-multipart